sudah buat exception utk sprint 1

+++ KouveePetShop/Jenis Hewan/CrudJenisHewan.xaml.cs
@@ -48,79 +48,91 @@ namespace KouveePetShop
 
         private void DataGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
-            DataGrid gd = (DataGrid)sender;
-            DataRowView selected_row = gd.SelectedItem as DataRowView;
-            if (selected_row != null)
+            try
             {
-                NamaJenisHewanText.Text = selected_row["NAMA_JENIS"].ToString();
-                IdJenisHewanText.Text = selected_row["ID_JENIS_HEWAN"].ToString();
+                DataGrid gd = (DataGrid)sender;
+                DataRowView selected_row = gd.SelectedItem as DataRowView;
+                if (selected_row != null)
+                {
+                    NamaJenisHewanText.Text = selected_row["NAMA_JENIS"].ToString();
+                    IdJenisHewanText.Text = selected_row["ID_JENIS_HEWAN"].ToString();
+                }
             }
+            catch(Exception err)
+            {
+                MessageBox.Show(err.Message);
+            }            
         }
 
         private void BtnTambah_Click(object sender, RoutedEventArgs e)
         {
-            /*
-            string jenishewan = NamaJenisHewanText.Text;
-
-            if(jenishewan == "")
+            try
             {
-                MessageBox.Show("Masukkan data pada field yang tersedia!");
-            }
-            else
-            {
-                bool a = tambahJenisHewan(jenishewan);
-                try
+                using (MySqlCommand cmd = new MySqlCommand())
                 {
-                    if (a)
+                    try
                     {
-                        MessageBox.Show("Berhasil Diinputkan!");
-                        conn.Close();
+                        conn.Open();
+                        cmd.CommandText = "INSERT INTO JENIS_HEWAN(NAMA_JENIS) VALUES(@jenis)";
+                        cmd.CommandType = CommandType.Text;
+                        cmd.Connection = conn;
+
+                        if(NamaJenisHewanText.Text == "")
+                        {
+                            MessageBox.Show("Field tidak boleh kosong", "Warning");
+                            conn.Close();
+                        }
+                        else
+                        {
+                            cmd.Parameters.AddWithValue("@jenis", NamaJenisHewanText.Text);
+                            cmd.ExecuteNonQuery();
+                            conn.Close();
+                            GetRecords();
+                            MessageBox.Show("Berhasil ditambahkan");
+                            NamaJenisHewanText.Clear();
+                        }
+                    }
+                    catch (Exception err)
+                    {
+                        MessageBox.Show(err.Message);
                     }
                 }
-                catch(Exception err)
-                {
-                    MessageBox.Show(err.Message);
-                }
-            }*/
-            using (MySqlCommand cmd = new MySqlCommand())
-            {
-                try
-                {
-                    conn.Open();
-                    cmd.CommandText = "INSERT INTO JENIS_HEWAN(NAMA_JENIS) VALUES(@jenis)";
-                    cmd.CommandType = CommandType.Text;
-                    cmd.Connection = conn;
-
-                    cmd.Parameters.AddWithValue("@jenis", NamaJenisHewanText.Text);
-
-
-                    cmd.ExecuteNonQuery();
-                    conn.Close();
-                    GetRecords();
-                    MessageBox.Show("Berhasil ditambahkan");
-                    NamaJenisHewanText.Clear();
-                    // conn.Close();
-                }
-                catch (Exception err)
-                {
-                    MessageBox.Show(err.Message);
-                }
             }
+            catch(Exception err)
+            {
+                MessageBox.Show(err.Message);
+            }           
         }
 
         private void BtnEdit_Click(object sender, RoutedEventArgs e)
         {
             try
             {
-                conn.Open();
-                ds = new DataSet();
-                adapter = new MySqlDataAdapter("update jenis_hewan set NAMA_JENIS = '" + NamaJenisHewanText.Text + "'where ID_JENIS_HEWAN = '" + IdJenisHewanText.Text + "'", conn);
-                adapter.Fill(ds, "jenis_hewan");
-                conn.Close();
-                GetRecords();
-                MessageBox.Show("Berhasil Diedit!");
-                NamaJenisHewanText.Clear();
-                IdJenisHewanText.Clear();
+                if(NamaJenisHewanText.Text == "")
+                {
+                    MessageBox.Show("Field tidak boleh kosong", "Warning");
+                }
+                else
+                {
+                    conn.Open();
+                    ds = new DataSet();
+                    adapter = new MySqlDataAdapter("update jenis_hewan set NAMA_JENIS = '" + NamaJenisHewanText.Text + "'where ID_JENIS_HEWAN = '" + IdJenisHewanText.Text + "'", conn);
+
+                    if (NamaJenisHewanText.Text == "")
+                    {
+                        MessageBox.Show("Field tidak boleh kosong", "Warning");
+                        conn.Close();
+                    }
+                    else
+                    {
+                        adapter.Fill(ds, "jenis_hewan");
+                        conn.Close();
+                        GetRecords();
+                        MessageBox.Show("Berhasil Diedit!");
+                        NamaJenisHewanText.Clear();
+                        IdJenisHewanText.Clear();
+                    }                    
+                }                
             }
             catch (Exception err)
             {
@@ -130,23 +142,43 @@ namespace KouveePetShop
 
         private void BtnHapus_Click(object sender, RoutedEventArgs e)
         {
-            conn.Open();
+            string message = "Apakah anda ingin menghapus data ini ?";
+            string caption = "Warning";
+            MessageBoxButton buttons = MessageBoxButton.YesNo;
+            MessageBoxImage icon = MessageBoxImage.Question;
             try
             {
-                using (MySqlCommand cmd = new MySqlCommand())
+                if(NamaJenisHewanText.Text == "")
                 {
-                    cmd.CommandText = "DELETE FROM JENIS_HEWAN WHERE ID_JENIS_HEWAN = @idjenis";
-                    cmd.CommandType = CommandType.Text;
-                    cmd.Connection = conn;
-
-                    cmd.Parameters.AddWithValue("@idjenis", IdJenisHewanText.Text);
-                    cmd.ExecuteNonQuery();
-                    conn.Close();
-                    GetRecords();
-                    MessageBox.Show("Berhasil Dihapus!");
-                    NamaJenisHewanText.Clear();
-                    IdJenisHewanText.Clear();
+                    MessageBox.Show("Silahkan pilih data terlebih dahulu", "Warning");
+                    return;
                 }
+                else
+                {
+                    using (MySqlCommand cmd = new MySqlCommand())
+                    {
+                        conn.Open();
+                        cmd.CommandText = "DELETE FROM JENIS_HEWAN WHERE ID_JENIS_HEWAN = @idjenis";
+                        cmd.CommandType = CommandType.Text;
+                        cmd.Connection = conn;
+
+                        if (MessageBox.Show(message, caption, buttons, icon) == MessageBoxResult.Yes)
+                        {
+                            cmd.Parameters.AddWithValue("@idjenis", IdJenisHewanText.Text);
+                            cmd.ExecuteNonQuery();
+                            conn.Close();
+                            GetRecords();
+                            MessageBox.Show("Berhasil Dihapus!");
+                            ClearData();
+                        }
+                        else
+                        {
+                            conn.Close();
+                            ClearData();
+                            return;
+                        }    
+                    }
+                }               
             }
             catch (Exception err)
             {
@@ -177,23 +209,57 @@ namespace KouveePetShop
 
         private void TextSearch_TextChanged(object sender, TextChangedEventArgs e)
         {
-            //Fungsi untuk mencari pegawai sesuai nama
-            DataTable dt = new DataTable();
-            //dt.Columns.Add(new DataColumn("Nama_Jenis"));
-            //DataView dv = new DataView(dt);
-            //dv.RowFilter = string.Format("Select * from jenis_hewan where Nama_Jenis LIKE '%{0}%'", NamaJenisHewanText.Text);
-            MySqlDataAdapter adp = new MySqlDataAdapter("Select ID_JENIS_HEWAN, NAMA_JENIS from jenis_hewan where Nama_Jenis LIKE '" + NamaJenisHewanText.Text + "%'", conn);
-            adp.Fill(dt);
-            //DataGrid.Items.Refresh();
-            DataGrid.DataContext = dt;
+            try
+            {
+                //Fungsi untuk mencari pegawai sesuai nama
+                DataTable dt = new DataTable();
+                //dt.Columns.Add(new DataColumn("Nama_Jenis"));
+                //DataView dv = new DataView(dt);
+                //dv.RowFilter = string.Format("Select * from jenis_hewan where Nama_Jenis LIKE '%{0}%'", NamaJenisHewanText.Text);
+                MySqlDataAdapter adp = new MySqlDataAdapter("Select ID_JENIS_HEWAN, NAMA_JENIS from jenis_hewan where Nama_Jenis LIKE '" + NamaJenisHewanText.Text + "%'", conn);
+                adp.Fill(dt);
+                //DataGrid.Items.Refresh();
+                DataGrid.DataContext = dt;
+            }
+            catch(Exception err)
+            {
+                MessageBox.Show(err.Message);
+            }           
         }
 
         private void BtnTampil_Click(object sender, RoutedEventArgs e)
         {
-            // Tampil data ke dataGrid
-            MySqlCommand cmd = new MySqlCommand("select ID_JENIS_HEWAN, NAMA_JENIS from jenis_hewan", conn);
             try
             {
+                // Tampil data ke dataGrid
+                MySqlCommand cmd = new MySqlCommand("select ID_JENIS_HEWAN, NAMA_JENIS from jenis_hewan", conn);
+                try
+                {
+                    conn.Open();
+                    DataTable dt = new DataTable();
+                    dt.Load(cmd.ExecuteReader());
+                    conn.Close();
+
+                    DataGrid.DataContext = dt;
+                    ClearData();
+                }
+                catch (MySqlException d)
+                {
+                    MessageBox.Show(d.Message);
+                }
+            }
+            catch(Exception err)
+            {
+                MessageBox.Show(err.Message);
+            }          
+        }
+
+        private void GetRecords()
+        {
+            try
+            {
+                MySqlCommand cmd = new MySqlCommand("select ID_JENIS_HEWAN, NAMA_JENIS from jenis_hewan", conn);
+                DataGrid.Items.Refresh();
                 conn.Open();
                 DataTable dt = new DataTable();
                 dt.Load(cmd.ExecuteReader());
@@ -201,22 +267,16 @@ namespace KouveePetShop
 
                 DataGrid.DataContext = dt;
             }
-            catch (MySqlException d)
+            catch(Exception err)
             {
-                MessageBox.Show(d.Message);
-            }
+                MessageBox.Show(err.Message);
+            }           
         }
 
-        private void GetRecords()
+        private void ClearData()
         {
-            MySqlCommand cmd = new MySqlCommand("select ID_JENIS_HEWAN, NAMA_JENIS from jenis_hewan", conn);
-            DataGrid.Items.Refresh();
-            conn.Open();
-            DataTable dt = new DataTable();
-            dt.Load(cmd.ExecuteReader());
-            conn.Close();
-
-            DataGrid.DataContext = dt;
+            NamaJenisHewanText.Clear();
+            IdJenisHewanText.Clear();
         }
     }
 }
